@@ -4,8 +4,8 @@ version: 0.3.0+1
 homepage: https://github.com/ferraridamiano/units_converter
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=2.14.0 <3.0.0'
 
 dev_dependencies:
-  pedantic: ^1.11.0
+  lints: ^1.0.1
   test: ^1.16.5
