@@ -1,6 +1,6 @@
-import 'Property.dart';
-import 'UtilsConversion.dart';
-import 'Unit.dart';
+import 'package:units_converter/Models/property.dart';
+import 'package:units_converter/Models/unit.dart';
+import 'package:units_converter/utils/utils_conversion.dart';
 
 //Available POWER units
 enum POWER {
@@ -9,8 +9,8 @@ enum POWER {
   kilowatt,
   megawatt,
   gigawatt,
-  european_horse_power,
-  imperial_horse_power,
+  europeanHorsePower,
+  imperialHorsePower,
 }
 
 class Power extends Property<POWER, double> {
@@ -21,8 +21,8 @@ class Power extends Property<POWER, double> {
     POWER.kilowatt: 'kW',
     POWER.megawatt: 'MW',
     POWER.gigawatt: 'GW',
-    POWER.european_horse_power: 'hp(M)',
-    POWER.imperial_horse_power: 'hp(I)',
+    POWER.europeanHorsePower: 'hp(M)',
+    POWER.imperialHorsePower: 'hp(I)',
   };
 
   int significantFigures;
@@ -36,9 +36,11 @@ class Power extends Property<POWER, double> {
   /// ```
   Power({this.significantFigures = 10, this.removeTrailingZeros = true, name}) {
     size = POWER.values.length;
-    this.name = name ?? PROPERTY.POWER;
-    POWER.values.forEach((element) => unitList.add(Unit(element, symbol: mapSymbols[element])));
-    unit_conversion = Node(name: POWER.watt, leafNodes: [
+    this.name = name ?? PROPERTY.power;
+    for (POWER val in POWER.values) {
+      unitList.add(Unit(val, symbol: mapSymbols[val]));
+    }
+    unitConversion = Node(name: POWER.watt, leafNodes: [
       Node(
         coefficientProduct: 1e-3,
         name: POWER.milliwatt,
@@ -57,11 +59,11 @@ class Power extends Property<POWER, double> {
       ),
       Node(
         coefficientProduct: 735.49875,
-        name: POWER.european_horse_power,
+        name: POWER.europeanHorsePower,
       ),
       Node(
         coefficientProduct: 745.69987158,
-        name: POWER.imperial_horse_power,
+        name: POWER.imperialHorsePower,
       ),
     ]);
   }
@@ -72,7 +74,7 @@ class Power extends Property<POWER, double> {
     super.convert(name, value);
     if (value == null) return;
     for (var i = 0; i < POWER.values.length; i++) {
-      unitList[i].value = unit_conversion.getByName(POWER.values.elementAt(i))?.value;
+      unitList[i].value = unitConversion.getByName(POWER.values.elementAt(i))?.value;
       unitList[i].stringValue = mantissaCorrection(unitList[i].value!, significantFigures, removeTrailingZeros);
     }
   }
@@ -82,6 +84,6 @@ class Power extends Property<POWER, double> {
   Unit get kilowatt => getUnit(POWER.kilowatt);
   Unit get megawatt => getUnit(POWER.megawatt);
   Unit get gigawatt => getUnit(POWER.gigawatt);
-  Unit get european_horse_power => getUnit(POWER.european_horse_power);
-  Unit get imperial_horse_power => getUnit(POWER.imperial_horse_power);
+  Unit get europeanHorsePower => getUnit(POWER.europeanHorsePower);
+  Unit get imperialHorsePower => getUnit(POWER.imperialHorsePower);
 }
