@@ -1,18 +1,18 @@
-import 'Property.dart';
-import 'UtilsConversion.dart';
-import 'Unit.dart';
+import 'package:units_converter/Models/property.dart';
+import 'package:units_converter/Models/unit.dart';
+import 'package:units_converter/utils/utils_conversion.dart';
 
 //Available SHOE_SIZE units
 enum SHOE_SIZE {
   centimeters,
   inches,
-  eu_china,
-  uk_india_child,
-  uk_india_man,
-  uk_india_woman,
-  usa_canada_child,
-  usa_canada_man,
-  usa_canada_woman,
+  euChina,
+  ukIndiaChild,
+  ukIndiaMan,
+  ukIndiaWoman,
+  usaCanadaChild,
+  usaCanadaMan,
+  usaCanadaWoman,
   japan,
 }
 
@@ -34,44 +34,46 @@ class ShoeSize extends Property<SHOE_SIZE, double> {
   /// ```
   ShoeSize({this.significantFigures = 10, this.removeTrailingZeros = true, name}) {
     size = SHOE_SIZE.values.length;
-    this.name = name ?? PROPERTY.SHOE_SIZE;
-    SHOE_SIZE.values.forEach((element) => unitList.add(Unit(element, symbol: mapSymbols[element])));
-    unit_conversion = Node(name: SHOE_SIZE.centimeters, leafNodes: [
+    this.name = name ?? PROPERTY.shoeSize;
+    for (SHOE_SIZE val in SHOE_SIZE.values) {
+      unitList.add(Unit(val, symbol: mapSymbols[val]));
+    }
+    unitConversion = Node(name: SHOE_SIZE.centimeters, leafNodes: [
       Node(
         coefficientProduct: 1 / 1.5,
         coefficientSum: -1.5,
-        name: SHOE_SIZE.eu_china,
+        name: SHOE_SIZE.euChina,
       ),
       Node(coefficientProduct: 2.54, name: SHOE_SIZE.inches, leafNodes: [
         Node(
           coefficientProduct: 1 / 3,
           coefficientSum: 10 / 3,
-          name: SHOE_SIZE.uk_india_child,
+          name: SHOE_SIZE.ukIndiaChild,
         ),
         Node(
           coefficientProduct: 1 / 3,
           coefficientSum: 23 / 3,
-          name: SHOE_SIZE.uk_india_man,
+          name: SHOE_SIZE.ukIndiaMan,
         ),
         Node(
           coefficientProduct: 1 / 3,
           coefficientSum: 23.5 / 3,
-          name: SHOE_SIZE.uk_india_woman,
+          name: SHOE_SIZE.ukIndiaWoman,
         ),
         Node(
           coefficientProduct: 1 / 3,
           coefficientSum: 49 / 9,
-          name: SHOE_SIZE.usa_canada_child,
+          name: SHOE_SIZE.usaCanadaChild,
         ),
         Node(
           coefficientProduct: 1 / 3,
           coefficientSum: 22 / 3,
-          name: SHOE_SIZE.usa_canada_man,
+          name: SHOE_SIZE.usaCanadaMan,
         ),
         Node(
           coefficientProduct: 1 / 3,
           coefficientSum: 21 / 3,
-          name: SHOE_SIZE.usa_canada_woman,
+          name: SHOE_SIZE.usaCanadaWoman,
         ),
       ]),
       Node(
@@ -87,19 +89,19 @@ class ShoeSize extends Property<SHOE_SIZE, double> {
     super.convert(name, value);
     if (value == null) return;
     for (var i = 0; i < SHOE_SIZE.values.length; i++) {
-      unitList[i].value = unit_conversion.getByName(SHOE_SIZE.values.elementAt(i))?.value;
+      unitList[i].value = unitConversion.getByName(SHOE_SIZE.values.elementAt(i))?.value;
       unitList[i].stringValue = mantissaCorrection(unitList[i].value!, significantFigures, removeTrailingZeros);
     }
   }
 
   Unit get centimeters => getUnit(SHOE_SIZE.centimeters);
   Unit get inches => getUnit(SHOE_SIZE.inches);
-  Unit get eu_china => getUnit(SHOE_SIZE.eu_china);
-  Unit get uk_india_child => getUnit(SHOE_SIZE.uk_india_child);
-  Unit get uk_india_man => getUnit(SHOE_SIZE.uk_india_man);
-  Unit get uk_india_woman => getUnit(SHOE_SIZE.uk_india_woman);
-  Unit get usa_canada_child => getUnit(SHOE_SIZE.usa_canada_child);
-  Unit get usa_canada_man => getUnit(SHOE_SIZE.usa_canada_man);
-  Unit get usa_canada_woman => getUnit(SHOE_SIZE.usa_canada_woman);
+  Unit get euChina => getUnit(SHOE_SIZE.euChina);
+  Unit get ukIndiaChild => getUnit(SHOE_SIZE.ukIndiaChild);
+  Unit get ukIndiaMan => getUnit(SHOE_SIZE.ukIndiaMan);
+  Unit get ukIndiaWoman => getUnit(SHOE_SIZE.ukIndiaWoman);
+  Unit get usaCanadaChild => getUnit(SHOE_SIZE.usaCanadaChild);
+  Unit get usaCanadaMan => getUnit(SHOE_SIZE.usaCanadaMan);
+  Unit get usaCanadaWoman => getUnit(SHOE_SIZE.usaCanadaWoman);
   Unit get japan => getUnit(SHOE_SIZE.japan);
 }
