@@ -1,24 +1,24 @@
-import 'Property.dart';
-import 'UtilsConversion.dart';
-import 'Unit.dart';
+import 'package:units_converter/Models/property.dart';
+import 'package:units_converter/Models/unit.dart';
+import 'package:units_converter/utils/utils_conversion.dart';
 
 //Available SPEED units
 enum SPEED {
-  meters_per_second,
-  kilometers_per_hour,
-  miles_per_hour,
+  metersPerSecond,
+  kilometersPerHour,
+  milesPerHour,
   knots,
-  feets_per_second,
+  feetsPerSecond,
 }
 
 class Speed extends Property<SPEED, double> {
   //Map between units and its symbol
   final Map<SPEED, String> mapSymbols = {
-    SPEED.meters_per_second: 'm/s',
-    SPEED.kilometers_per_hour: 'km/h',
-    SPEED.miles_per_hour: 'mi/h',
+    SPEED.metersPerSecond: 'm/s',
+    SPEED.kilometersPerHour: 'km/h',
+    SPEED.milesPerHour: 'mi/h',
     SPEED.knots: 'kts',
-    SPEED.feets_per_second: 'ft/s',
+    SPEED.feetsPerSecond: 'ft/s',
   };
 
   int significantFigures;
@@ -32,13 +32,15 @@ class Speed extends Property<SPEED, double> {
   /// ```
   Speed({this.significantFigures = 10, this.removeTrailingZeros = true, name}) {
     size = SPEED.values.length;
-    this.name = name ?? PROPERTY.SPEED;
-    SPEED.values.forEach((element) => unitList.add(Unit(element, symbol: mapSymbols[element])));
-    unit_conversion = Node(name: SPEED.meters_per_second, leafNodes: [
-      Node(coefficientProduct: 1 / 3.6, name: SPEED.kilometers_per_hour, leafNodes: [
+    this.name = name ?? PROPERTY.speed;
+    for (SPEED val in SPEED.values) {
+      unitList.add(Unit(val, symbol: mapSymbols[val]));
+    }
+    unitConversion = Node(name: SPEED.metersPerSecond, leafNodes: [
+      Node(coefficientProduct: 1 / 3.6, name: SPEED.kilometersPerHour, leafNodes: [
         Node(
           coefficientProduct: 1.609344,
-          name: SPEED.miles_per_hour,
+          name: SPEED.milesPerHour,
         ),
         Node(
           coefficientProduct: 1.852,
@@ -47,7 +49,7 @@ class Speed extends Property<SPEED, double> {
       ]),
       Node(
         coefficientProduct: 0.3048,
-        name: SPEED.feets_per_second,
+        name: SPEED.feetsPerSecond,
       ),
     ]);
   }
@@ -58,14 +60,14 @@ class Speed extends Property<SPEED, double> {
     super.convert(name, value);
     if (value == null) return;
     for (var i = 0; i < SPEED.values.length; i++) {
-      unitList[i].value = unit_conversion.getByName(SPEED.values.elementAt(i))?.value;
+      unitList[i].value = unitConversion.getByName(SPEED.values.elementAt(i))?.value;
       unitList[i].stringValue = mantissaCorrection(unitList[i].value!, significantFigures, removeTrailingZeros);
     }
   }
 
-  Unit get meters_per_second => getUnit(SPEED.meters_per_second);
-  Unit get kilometers_per_hour => getUnit(SPEED.kilometers_per_hour);
-  Unit get miles_per_hour => getUnit(SPEED.miles_per_hour);
+  Unit get metersPerSecond => getUnit(SPEED.metersPerSecond);
+  Unit get kilometersPerHour => getUnit(SPEED.kilometersPerHour);
+  Unit get milesPerHour => getUnit(SPEED.milesPerHour);
   Unit get knots => getUnit(SPEED.knots);
-  Unit get feets_per_second => getUnit(SPEED.feets_per_second);
+  Unit get feetsPerSecond => getUnit(SPEED.feetsPerSecond);
 }
