@@ -1,24 +1,24 @@
-import 'Property.dart';
-import 'UtilsConversion.dart';
-import 'Unit.dart';
+import 'package:units_converter/Models/property.dart';
+import 'package:units_converter/Models/unit.dart';
+import 'package:units_converter/utils/utils_conversion.dart';
 
 //Available TORQUE units
 enum TORQUE {
-  newton_meter,
-  dyne_meter,
-  pound_force_feet,
-  kilogram_force_meter,
-  poundal_meter,
+  newtonMeter,
+  dyneMeter,
+  poundForceFeet,
+  kilogramForceMeter,
+  poundalMeter,
 }
 
 class Torque extends Property<TORQUE, double> {
   //Map between units and its symbol
   final Map<TORQUE, String> mapSymbols = {
-    TORQUE.newton_meter: 'N·m',
-    TORQUE.dyne_meter: 'dyn·m',
-    TORQUE.pound_force_feet: 'lbf·ft',
-    TORQUE.kilogram_force_meter: 'kgf·m',
-    TORQUE.poundal_meter: 'pdl·m',
+    TORQUE.newtonMeter: 'N·m',
+    TORQUE.dyneMeter: 'dyn·m',
+    TORQUE.poundForceFeet: 'lbf·ft',
+    TORQUE.kilogramForceMeter: 'kgf·m',
+    TORQUE.poundalMeter: 'pdl·m',
   };
 
   int significantFigures;
@@ -32,24 +32,26 @@ class Torque extends Property<TORQUE, double> {
   /// ```
   Torque({this.significantFigures = 10, this.removeTrailingZeros = true, name}) {
     size = TORQUE.values.length;
-    this.name = name ?? PROPERTY.TORQUE;
-    TORQUE.values.forEach((element) => unitList.add(Unit(element, symbol: mapSymbols[element])));
-    unit_conversion = Node(name: TORQUE.newton_meter, leafNodes: [
+    this.name = name ?? PROPERTY.torque;
+    for (TORQUE val in TORQUE.values) {
+      unitList.add(Unit(val, symbol: mapSymbols[val]));
+    }
+    unitConversion = Node(name: TORQUE.newtonMeter, leafNodes: [
       Node(
         coefficientProduct: 1e-5,
-        name: TORQUE.dyne_meter,
+        name: TORQUE.dyneMeter,
       ),
       Node(
         coefficientProduct: 1.35581794902490555,
-        name: TORQUE.pound_force_feet,
+        name: TORQUE.poundForceFeet,
       ),
       Node(
         coefficientProduct: 9.807,
-        name: TORQUE.kilogram_force_meter,
+        name: TORQUE.kilogramForceMeter,
       ),
       Node(
         coefficientProduct: 0.138254954376,
-        name: TORQUE.poundal_meter,
+        name: TORQUE.poundalMeter,
       ),
     ]);
   }
@@ -60,14 +62,14 @@ class Torque extends Property<TORQUE, double> {
     super.convert(name, value);
     if (value == null) return;
     for (var i = 0; i < TORQUE.values.length; i++) {
-      unitList[i].value = unit_conversion.getByName(TORQUE.values.elementAt(i))?.value;
+      unitList[i].value = unitConversion.getByName(TORQUE.values.elementAt(i))?.value;
       unitList[i].stringValue = mantissaCorrection(unitList[i].value!, significantFigures, removeTrailingZeros);
     }
   }
 
-  Unit get newton_meter => getUnit(TORQUE.newton_meter);
-  Unit get dyne_meter => getUnit(TORQUE.dyne_meter);
-  Unit get pound_force_feet => getUnit(TORQUE.pound_force_feet);
-  Unit get kilogram_force_meter => getUnit(TORQUE.kilogram_force_meter);
-  Unit get poundal_meter => getUnit(TORQUE.poundal_meter);
+  Unit get newtonMeter => getUnit(TORQUE.newtonMeter);
+  Unit get dyneMeter => getUnit(TORQUE.dyneMeter);
+  Unit get poundForceFeet => getUnit(TORQUE.poundForceFeet);
+  Unit get kilogramForceMeter => getUnit(TORQUE.kilogramForceMeter);
+  Unit get poundalMeter => getUnit(TORQUE.poundalMeter);
 }
